Improve file layout and error handling
