+++ cmd/kubejob/main.go
@@ -10,7 +10,6 @@ import (
 
 	"github.com/goccy/kubejob"
 	"github.com/jessevdk/go-flags"
-	"golang.org/x/xerrors"
 	_ "k8s.io/client-go/plugin/pkg/client/auth"
 	"k8s.io/client-go/rest"
 	"k8s.io/client-go/tools/clientcmd"
@@ -38,7 +37,7 @@ func getKubeConfig() string {
 func loadConfig() (*rest.Config, error) {
 	cfg, err := clientcmd.BuildConfigFromFlags("", getKubeConfig())
 	if err != nil {
-		return nil, xerrors.Errorf("failed to create config: %w", err)
+		return nil, fmt.Errorf("failed to create config: %w", err)
 	}
 	return cfg, nil
 }
@@ -50,44 +49,44 @@ func namespace(opt option) (string, error) {
 	rules := &clientcmd.ClientConfigLoadingRules{ExplicitPath: getKubeConfig()}
 	c, err := rules.Load()
 	if err != nil {
-		return "", xerrors.Errorf("failed to load default namespace: %w", err)
+		return "", fmt.Errorf("failed to load default namespace: %w", err)
 	}
 	return c.Contexts[c.CurrentContext].Namespace, nil
 }
 
 func _main(args []string, opt option) error {
 	if opt.Image == "" && opt.File == "" {
-		return xerrors.Errorf("image or file option must be specified")
+		return fmt.Errorf("image or file option must be specified")
 	}
 	cfg, err := loadConfig()
 	if err != nil {
-		return xerrors.Errorf("failed to load config: %w", err)
+		return fmt.Errorf("failed to load config: %w", err)
 	}
 	ns, err := namespace(opt)
 	if err != nil {
-		return xerrors.Errorf("failed to get namespace: %w", err)
+		return fmt.Errorf("failed to get namespace: %w", err)
 	}
 	var job *kubejob.Job
 	if opt.File != "" {
 		file, err := os.Open(opt.File)
 		if err != nil {
-			return xerrors.Errorf("failed to open file %s: %w", opt.File, err)
+			return fmt.Errorf("failed to open file %s: %w", opt.File, err)
 		}
 		j, err := kubejob.NewJobBuilder(cfg, ns).BuildWithReader(file)
 		if err != nil {
-			return xerrors.Errorf("failed to build job: %w", err)
+			return err
 		}
 		job = j
 	} else {
 		if len(args) == 0 {
-			return xerrors.Errorf("command is required. please speficy after '--' section")
+			return fmt.Errorf("command is required. please speficy after '--' section")
 		}
 		j, err := kubejob.NewJobBuilder(cfg, ns).
 			SetImage(opt.Image).
 			SetCommand(args).
 			Build()
 		if err != nil {
-			return xerrors.Errorf("failed to build job: %w", err)
+			return err
 		}
 		job = j
 	}
@@ -104,7 +103,7 @@ func _main(args []string, opt option) error {
 	}()
 
 	if err := job.Run(ctx); err != nil {
-		return xerrors.Errorf("failed to run job: %w", err)
+		return err
 	}
 	return nil
 }
