+++ copy.go
@@ -10,23 +10,18 @@ import (
 	"path/filepath"
 	"strings"
 
-	"golang.org/x/xerrors"
 	core "k8s.io/api/core/v1"
 	"k8s.io/client-go/kubernetes/scheme"
 	"k8s.io/client-go/tools/remotecommand"
 )
 
-var (
-	ErrEmptyFileSpecified = xerrors.New("cannot specify empty file path")
-)
-
 // CopyToPod copy directory or files to specified path on Pod.
 func (e *JobExecutor) CopyToPod(srcPath, dstPath string) error {
 	if len(srcPath) == 0 || len(dstPath) == 0 {
-		return ErrEmptyFileSpecified
+		return errCopyWithEmptyPath(srcPath, dstPath)
 	}
 	if _, err := os.Stat(srcPath); err != nil {
-		return xerrors.Errorf("%s doesn't exist in local filesystem", srcPath)
+		return errCopy(srcPath, dstPath, fmt.Errorf("%s doesn't exist in local filesystem", srcPath))
 	}
 
 	// trim slash as the last character
@@ -61,7 +56,7 @@ func (e *JobExecutor) CopyToPod(srcPath, dstPath string) error {
 	url := req.URL()
 	exec, err := remotecommand.NewSPDYExecutor(e.job.config, "POST", url)
 	if err != nil {
-		return xerrors.Errorf("failed to create spdy executor: %w", err)
+		return fmt.Errorf("job: failed to create spdy executor: %w", err)
 	}
 
 	reader, writer := io.Pipe()
@@ -73,19 +68,14 @@ func (e *JobExecutor) CopyToPod(srcPath, dstPath string) error {
 	}()
 
 	var outCapturer bytes.Buffer
-	if err := exec.Stream(remotecommand.StreamOptions{
+	readerErr := exec.Stream(remotecommand.StreamOptions{
 		Stdin:  reader,
 		Stdout: &outCapturer,
 		Stderr: &outCapturer,
 		Tty:    false,
-	}); err != nil {
-		if writerErr != nil {
-			return xerrors.Errorf("failed to copy %s. %s: %w", outCapturer.String(), writerErr, err)
-		}
-		return xerrors.Errorf("failed to copy %s: %w", outCapturer.String(), err)
-	}
-	if writerErr != nil {
-		return xerrors.Errorf("failed to write with tar: %w", writerErr)
+	})
+	if readerErr != nil || writerErr != nil {
+		return errCopyWithReaderWriter(srcPath, dstPath, readerErr, writerErr, outCapturer.String())
 	}
 	return nil
 }
@@ -93,7 +83,7 @@ func (e *JobExecutor) CopyToPod(srcPath, dstPath string) error {
 // CopyFromPod copy directory or files from specified path on Pod.
 func (e *JobExecutor) CopyFromPod(srcPath, dstPath string) error {
 	if len(srcPath) == 0 || len(dstPath) == 0 {
-		return ErrEmptyFileSpecified
+		return errCopyWithEmptyPath(srcPath, dstPath)
 	}
 
 	pod := e.Pod
@@ -112,7 +102,7 @@ func (e *JobExecutor) CopyFromPod(srcPath, dstPath string) error {
 	url := req.URL()
 	exec, err := remotecommand.NewSPDYExecutor(e.job.config, "POST", url)
 	if err != nil {
-		return xerrors.Errorf("failed to create spdy executor: %w", err)
+		return fmt.Errorf("job: failed to create spdy executor: %w", err)
 	}
 	reader, writer := io.Pipe()
 
@@ -135,15 +125,9 @@ func (e *JobExecutor) CopyFromPod(srcPath, dstPath string) error {
 	// tar trims the leading '/' if it's there
 	tarPrefix := strings.TrimLeft(srcPath, "/")
 	tarPrefix = e.trimShortcutPath(path.Clean(tarPrefix))
-	if err := e.untarAll(reader, &errCapturer, tarPrefix, srcPath, dstPath); err != nil {
-		return xerrors.Errorf("failed to untar: %w", err)
-	}
-	if writerErr != nil {
-		err := errCapturer.String()
-		if len(err) > 0 {
-			return xerrors.Errorf("failed to write %s: %w", err, writerErr)
-		}
-		return xerrors.Errorf("failed to write: %w", writerErr)
+	readerErr := e.untarAll(reader, &errCapturer, tarPrefix, srcPath, dstPath)
+	if readerErr != nil || writerErr != nil {
+		return errCopyWithReaderWriter(srcPath, dstPath, readerErr, writerErr, errCapturer.String())
 	}
 	return nil
 }
@@ -183,7 +167,7 @@ func (e *JobExecutor) writeWithTar(w io.Writer, srcPath, dstPath string) error {
 		path.Dir(dstPath),
 		path.Base(dstPath),
 	); err != nil {
-		return xerrors.Errorf("failed to write recursive with tar: %w", err)
+		return err
 	}
 	return nil
 }
@@ -192,23 +176,23 @@ func (e *JobExecutor) writeRecursiveWithTar(w *tar.Writer, srcBase, srcFile, dst
 	srcPath := path.Join(srcBase, srcFile)
 	matchedPaths, err := filepath.Glob(srcPath)
 	if err != nil {
-		return xerrors.Errorf("failed to glob from %s: %w", srcPath, err)
+		return fmt.Errorf("failed to glob from %s: %w", srcPath, err)
 	}
 	for _, fpath := range matchedPaths {
 		stat, err := os.Lstat(fpath)
 		if err != nil {
-			return xerrors.Errorf("failed to lstat for %s: %w", fpath, err)
+			return fmt.Errorf("failed to lstat for %s: %w", fpath, err)
 		}
 		if stat.IsDir() {
 			entries, err := os.ReadDir(fpath)
 			if err != nil {
-				return xerrors.Errorf("failed to readdir %s: %w", fpath, err)
+				return fmt.Errorf("failed to readdir %s: %w", fpath, err)
 			}
 			if len(entries) == 0 {
 				hdr, _ := tar.FileInfoHeader(stat, fpath)
 				hdr.Name = dstFile
 				if err := w.WriteHeader(hdr); err != nil {
-					return xerrors.Errorf("failed to write header: %w", err)
+					return fmt.Errorf("failed to write header: %w", err)
 				}
 			}
 			for _, entry := range entries {
@@ -219,7 +203,7 @@ func (e *JobExecutor) writeRecursiveWithTar(w *tar.Writer, srcBase, srcFile, dst
 					dstBase,
 					path.Join(dstFile, entry.Name()),
 				); err != nil {
-					return xerrors.Errorf("failed to write recursive with tar for %s: %w", entry.Name(), err)
+					return fmt.Errorf("failed to write recursive with tar for %s: %w", entry.Name(), err)
 				}
 			}
 			return nil
@@ -228,34 +212,34 @@ func (e *JobExecutor) writeRecursiveWithTar(w *tar.Writer, srcBase, srcFile, dst
 			hdr, _ := tar.FileInfoHeader(stat, fpath)
 			target, err := os.Readlink(fpath)
 			if err != nil {
-				return xerrors.Errorf("failed to readlink %s: %w", fpath, err)
+				return fmt.Errorf("failed to readlink %s: %w", fpath, err)
 			}
 
 			hdr.Linkname = target
 			hdr.Name = dstFile
 			if err := w.WriteHeader(hdr); err != nil {
-				return xerrors.Errorf("failed to write header: %w", err)
+				return fmt.Errorf("failed to write header: %w", err)
 			}
 		} else {
 			// regular file or other file type like pipe
 			hdr, err := tar.FileInfoHeader(stat, fpath)
 			if err != nil {
-				return xerrors.Errorf("failed to get header from %s: %w", fpath, err)
+				return fmt.Errorf("failed to get header from %s: %w", fpath, err)
 			}
 			hdr.Name = dstFile
 
 			if err := w.WriteHeader(hdr); err != nil {
-				return xerrors.Errorf("failed to write header: %w", err)
+				return fmt.Errorf("failed to write header: %w", err)
 			}
 
 			f, err := os.Open(fpath)
 			if err != nil {
-				return xerrors.Errorf("failed to open %s: %w", fpath, err)
+				return fmt.Errorf("failed to open %s: %w", fpath, err)
 			}
 			defer f.Close()
 
 			if _, err := io.Copy(w, f); err != nil {
-				return xerrors.Errorf("failed to copy %s: %w", fpath, err)
+				return fmt.Errorf("failed to copy %s: %w", fpath, err)
 			}
 			return nil
 		}
@@ -269,7 +253,7 @@ func (e *JobExecutor) untarAll(r io.Reader, errCapturer io.Writer, prefix, srcPa
 		header, err := tarReader.Next()
 		if err != nil {
 			if err != io.EOF {
-				return xerrors.Errorf("failed to get next header: %w", err)
+				return fmt.Errorf("failed to get next header: %w", err)
 			}
 			break
 		}
@@ -280,7 +264,7 @@ func (e *JobExecutor) untarAll(r io.Reader, errCapturer io.Writer, prefix, srcPa
 		// For the case where prefix is empty we need to ensure that the path
 		// is not absolute, which also indicates the tar file was tempered with.
 		if !strings.HasPrefix(header.Name, prefix) {
-			return xerrors.Errorf("tar contents corrupted")
+			return fmt.Errorf("tar contents corrupted")
 		}
 
 		// basic file information
@@ -294,11 +278,11 @@ func (e *JobExecutor) untarAll(r io.Reader, errCapturer io.Writer, prefix, srcPa
 
 		baseName := filepath.Dir(dstFileName)
 		if err := os.MkdirAll(baseName, 0755); err != nil {
-			return xerrors.Errorf("failed to mkdir %s: %w", baseName, err)
+			return fmt.Errorf("failed to mkdir %s: %w", baseName, err)
 		}
 		if header.FileInfo().IsDir() {
 			if err := os.MkdirAll(dstFileName, 0755); err != nil {
-				return xerrors.Errorf("failed to mkdir %s: %w", dstFileName, err)
+				return fmt.Errorf("failed to mkdir %s: %w", dstFileName, err)
 			}
 			continue
 		}
@@ -308,7 +292,7 @@ func (e *JobExecutor) untarAll(r io.Reader, errCapturer io.Writer, prefix, srcPa
 			continue
 		}
 		if err := e.copyFileFromReader(dstFileName, tarReader); err != nil {
-			return xerrors.Errorf("failed to copy file from reader %s: %w", dstFileName, err)
+			return fmt.Errorf("failed to copy file from reader %s: %w", dstFileName, err)
 		}
 	}
 
@@ -318,11 +302,11 @@ func (e *JobExecutor) untarAll(r io.Reader, errCapturer io.Writer, prefix, srcPa
 func (e *JobExecutor) copyFileFromReader(file string, reader io.Reader) error {
 	f, err := os.Create(file)
 	if err != nil {
-		return xerrors.Errorf("failed to create file %s: %w", file, err)
+		return err
 	}
 	defer f.Close()
 	if _, err := io.Copy(f, reader); err != nil {
-		return xerrors.Errorf("failed to copy %s: %w", file, err)
+		return err
 	}
 	return nil
 }
